Float / Int / whatever

+++ src/octoprint/printer.py
@@ -282,8 +282,8 @@ class Printer():
 		if self._selectedFile is None:
 			return
 		#Add M206 Z-offset command at start op print
-		zOffset = settings().getInt(["printerParameters", "zOffset"])
-		self.commands("M206 X0.0 Y0.0 Z%d" % zOffset)
+		zOffset = settings().getFloat(["printerParameters", "zOffset"])
+		self.commands("M206 X0.0 Y0.0 Z%f" % zOffset)
 		self._setCurrentZ(None)
 		self._comm.startPrint()
